assignment stack added
